requirements.txt
opencv-python==4.2.0.32
easydict==1.9
flask-restful==0.3.8
requests==2.24.0
Pillow==7.0.0
torch==1.6.0
torchvision==0.7.0
PyYAML==5.3.1
whl/torch-1.6.0-cp36-cp36m-manylinux1_x86_64.whl